requirements.txt
fastapi
uvicorn[standard]
streamlit
google-cloud-aiplatform 
python-dotenv
requests
python-decouple
pydantic
